requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.23.5